refactor(tarko): migrate FileResultRenderer out of generic components
Move FileResultRenderer from generic/components to standalone renderer
to eliminate coupling and establish clear architectural boundaries
